FIX: atan to atanf

+++ sources/drawing/draw_3d.c
@@ -6,7 +6,7 @@
 /*   By: jmoutous <jmoutous@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/06 14:20:36 by jmoutous          #+#    #+#             */
-/*   Updated: 2023/08/29 17:58:36 by jmoutous         ###   ########lyon.fr   */
+/*   Updated: 2023/08/29 18:36:47 by jmoutous         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -84,7 +84,7 @@ void	ft_draw_3d(t_data *game)
 	{
 		point = ft_3d_wall_collision(game, angle);
 		ft_draw_wall(game, point, i);
-		angle = atan((i + 0.0001 - WIN_WIDTH / 2) / game->screen_dst)
+		angle = atanf((i + 0.0001 - WIN_WIDTH / 2) / game->screen_dst)
 			+ game->degree + 0.0001;
 		i++;
 	}
